Move NoopMeter to package private
