+++ api/metrics/src/test/java/io/opentelemetry/api/metrics/NoopMeterProviderTest.java
@@ -3,9 +3,8 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.api.metrics.internal;
+package io.opentelemetry.api.metrics;
 
-import io.opentelemetry.api.metrics.MeterProvider;
 import org.junit.jupiter.api.Test;
 
 public class NoopMeterProviderTest {
